[react] Move addons and mixin tests to their packages
