+++ types/react-addons-linked-state-mixin/package.json
@@ -9,6 +9,8 @@
         "@types/react": "*"
     },
     "devDependencies": {
+        "@types/create-react-class": "*",
+        "@types/react-dom-factories": "*",
         "@types/react-addons-linked-state-mixin": "workspace:."
     },
     "owners": [
