+++ types/react-addons-linked-state-mixin/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "files": [
-        "index.d.ts"
+        "index.d.ts",
+        "react-addons-linked-state-mixin-tests.ts"
     ],
     "compilerOptions": {
         "module": "commonjs",
