+++ types/react-addons-pure-render-mixin/react-addons-pure-render-mixin-tests.ts
@@ -1,0 +1,11 @@
+
+import * as PureRenderMixin from "react-addons-pure-render-mixin";
+import createReactClass = require("create-react-class");
+import * as DOM from "react-dom-factories";
+
+createReactClass({
+    mixins: [PureRenderMixin],
+    render() {
+        return DOM.div(null);
+    },
+});
