+++ types/react-addons-pure-render-mixin/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "files": [
-        "index.d.ts"
+        "index.d.ts",
+        "react-addons-pure-render-mixin-tests.ts"
     ],
     "compilerOptions": {
         "module": "commonjs",
