+++ types/react/ts5.0/test/index.ts
@@ -2,12 +2,6 @@ import * as PropTypes from "prop-types";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import * as ReactDOMServer from "react-dom/server";
-import createFragment = require("react-addons-create-fragment");
-import * as LinkedStateMixin from "react-addons-linked-state-mixin";
-import * as PureRenderMixin from "react-addons-pure-render-mixin";
-import shallowCompare = require("react-addons-shallow-compare");
-import update = require("react-addons-update");
-import createReactClass = require("create-react-class");
 import * as DOM from "react-dom-factories";
 import "trusted-types";
 
@@ -180,10 +174,6 @@ class ModernComponent extends React.Component<Props, State, Snapshot>
                 onChange: event => console.log(event.target),
             }),
         );
-    }
-
-    shouldComponentUpdate(nextProps: Props, nextState: State, nextContext: any): boolean {
-        return shallowCompare(this, nextProps, nextState);
     }
 
     getSnapshotBeforeUpdate(prevProps: Readonly<Props>) {
@@ -612,75 +602,6 @@ const mappedChildrenArray6 = React.Children.map(renderPropsChildren, element => 
 // The return type may not be an array
 // @ts-expect-error
 const mappedChildrenArray7 = React.Children.map(nodeChildren, node => node).map;
-
-//
-// createFragment addon
-// --------------------------------------------------------------------------
-createFragment({
-    a: DOM.div(),
-    b: ["a", false, React.createElement("span")],
-});
-
-//
-// LinkedStateMixin addon
-// --------------------------------------------------------------------------
-createReactClass({
-    mixins: [LinkedStateMixin],
-    getInitialState() {
-        return {
-            isChecked: false,
-            message: "hello!",
-        };
-    },
-    render() {
-        return DOM.div(
-            null,
-            DOM.input({
-                type: "checkbox",
-                checkedLink: this.linkState("isChecked"),
-            }),
-            DOM.input({
-                type: "text",
-                valueLink: this.linkState("message"),
-            }),
-        );
-    },
-});
-
-//
-// PureRenderMixin addon
-// --------------------------------------------------------------------------
-createReactClass({
-    mixins: [PureRenderMixin],
-    render() {
-        return DOM.div(null);
-    },
-});
-
-//
-// update addon
-// --------------------------------------------------------------------------
-{
-    // These are copied from https://facebook.github.io/react/docs/update.html
-    const initialArray = [1, 2, 3];
-    const newArray = update(initialArray, { $push: [4] }); // => [1, 2, 3, 4]
-
-    const collection = [1, 2, { a: [12, 17, 15] }];
-    const newCollection = update(collection, { 2: { a: { $splice: [[1, 1, 13, 14]] } } });
-    // => [1, 2, {a: [12, 13, 14, 15]}]
-
-    const obj = { a: 5, b: 3 };
-    const newObj = update(obj, {
-        b: {
-            $apply: x => x * 2,
-        },
-    });
-    // => {a: 5, b: 6}
-    const newObj2 = update(obj, { b: { $set: obj.b * 2 } });
-
-    const objShallow = { a: 5, b: 3 };
-    const newObjShallow = update(obj, { $merge: { b: 6, c: 7 } }); // => {a: 5, b: 6, c: 7}
-}
 
 //
 // Events
